Refactor and reorganize ember-cli-build.js
